tmp/pi_to_s3/requirements.txt
boto3==1.34.39
botocore==1.34.39
jmespath==1.0.1
mysql-connector-python==8.3.0
pretty==0.1
python-dateutil==2.8.2
pytz==2024.1
s3transfer==0.10.0
six==1.16.0
urllib3==1.26.18
